// src/app/interfaces/data.inerface.ts
export interface IData {
    name?: string;
    symbol?: string;
    number?: string;
    atomic_mass?: string;
    density?: string;
    summary?: string;
    xpos?: string;
    ypos?: string;
    shells?: string;
    electron_configuration?: string;
    melt?: string;
    boil?: string;
    category?: string;
    phase?: string;
    appearance?: string;
    color?: string;
    discovered_by?: string;
}